requirements.txt
django~=5.0
django-environ
django-modeltranslation
pillow
gunicorn
whitenoise
